libuser: mem module
