+++ libuser/src/lib.rs
@@ -29,7 +29,6 @@ extern crate bitfield;
 
 #[macro_use]
 extern crate kfs_libutils;
-use kfs_libkern;
 #[macro_use]
 extern crate failure;
 
@@ -41,6 +40,7 @@ extern crate lazy_static;
 
 pub mod caps;
 pub mod syscalls;
+pub mod mem;
 pub mod types;
 pub mod ipc;
 pub mod sm;
@@ -54,8 +54,7 @@ mod log_impl;
 pub use kfs_libutils::io;
 
 use kfs_libutils as utils;
-use crate::error::{Error, LibuserError};
- 
+
 // TODO: report #[cfg(not(test))] and #[global_allocator]
 // BODY: `#[cfg(not(test))]` still compiles this item with cargo test,
 // BODY: but `#[cfg(target_os = "none")] does not. I think this is a bug,
@@ -65,34 +64,6 @@ use crate::error::{Error, LibuserError};
 #[cfg(target_os = "none")]
 #[global_allocator]
 static ALLOCATOR: allocator::Allocator = allocator::Allocator::new();
-
-/// Finds a free memory zone of the given size and alignment in the current
-/// process's virtual address space. Note that the address space is not reserved,
-/// a call to map_memory to that address space might fail if another thread
-/// maps to it first. It is recommended to use this function and the map syscall
-/// in a loop.
-///
-/// # Panics
-///
-/// Panics on underflow when size = 0.
-///
-/// Panics on underflow when align = 0.
-pub fn find_free_address(size: usize, align: usize) -> Result<usize, Error> {
-    let mut addr = 0;
-    // Go over the address space.
-    loop {
-        let (meminfo, _) = syscalls::query_memory(addr)?;
-        if meminfo.memtype == kfs_libkern::MemoryType::Unmapped {
-            let alignedbaseaddr = kfs_libutils::align_up_checked(meminfo.baseaddr, align).ok_or(LibuserError::AddressSpaceExhausted)?;
-
-            let alignment = alignedbaseaddr - meminfo.baseaddr;
-            if alignment.checked_add(size - 1).ok_or(LibuserError::AddressSpaceExhausted)? < meminfo.size {
-                return Ok(alignedbaseaddr)
-            }
-        }
-        addr = meminfo.baseaddr.checked_add(meminfo.size).ok_or(LibuserError::AddressSpaceExhausted)?;
-    }
-}
 
 // Runtime functions
 //
